+++ internal/handlers/category_handler.go
@@ -6,6 +6,7 @@ import (
 	"book-manager/internal/dto/common"
 	"book-manager/internal/services"
 	"book-manager/internal/utils"
+	"book-manager/internal/utils/auth"
 	"book-manager/internal/utils/enums/error_codes"
 
 	"github.com/labstack/echo/v4"
@@ -32,7 +33,8 @@ func (h *CategoryHandler) Create(c echo.Context) error {
 		return c.JSON(constants.StatusBadRequest, resp)
 	}
 
-	resp := h.Service.Create(&reqDto)
+	user := auth.GetCurrentUser(c)
+	resp := h.Service.Create(&reqDto, user)
 	return c.JSON(constants.StatusOK, resp)
 }
 
@@ -76,7 +78,8 @@ func (h *CategoryHandler) Update(c echo.Context) error {
 		return c.JSON(constants.StatusBadRequest, resp)
 	}
 
-	resp := h.Service.Update(&reqDto)
+	user := auth.GetCurrentUser(c)
+	resp := h.Service.Update(&reqDto, user)
 	return c.JSON(constants.StatusOK, resp)
 }
 
@@ -93,6 +96,7 @@ func (h *CategoryHandler) Delete(c echo.Context) error {
 		return c.JSON(constants.StatusBadRequest, resp)
 	}
 
-	resp := h.Service.Delete(&reqDto)
+	user := auth.GetCurrentUser(c)
+	resp := h.Service.Delete(&reqDto, user)
 	return c.JSON(constants.StatusOK, resp)
 }