+++ internal/models/user/role.go
@@ -1,1 +1,3 @@
 package user
+
+// TODO: Implement role model for user authorization later