+++ internal/models/user/user.go
@@ -1,1 +1,3 @@
 package user
+
+// TODO: Implement user model for authentication and authorization later